♻️refactor : EmotionUploadResponse -> EmotionRelatedPhotoResponse 변경 및 폴더 이동으로 인한 import 수정
