+++ emopic/src/main/java/mmm/emopic/app/domain/emotion/EmotionController.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import mmm.emopic.app.base.Dto.BaseResponse;
 import mmm.emopic.app.domain.emotion.dto.request.EmotionUploadRequest;
 import mmm.emopic.app.domain.emotion.dto.response.EmotionResponse;
-import mmm.emopic.app.domain.emotion.dto.response.EmotionUploadResponse;
+import mmm.emopic.app.domain.emotion.dto.response.EmotionRelatedPhotoResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -27,7 +27,7 @@ public class EmotionController {
 
     @PostMapping("/photos/{photoId}/emotions")
     public ResponseEntity<BaseResponse> saveEmotionsInPhoto(@Validated @RequestBody EmotionUploadRequest emotionUploadRequest, @PathVariable(name ="photoId") Long photoId){
-        EmotionUploadResponse response = emotionService.saveEmotionsInPhoto(emotionUploadRequest,photoId);
+        EmotionRelatedPhotoResponse response = emotionService.saveEmotionsInPhoto(emotionUploadRequest,photoId);
         return ResponseEntity.ok(new BaseResponse( HttpStatus.OK.value(), "감정 저장 성공",response));
     }
 }
