--- emopic/src/main/java/mmm/emopic/app/domain/emotion/PhotoEmotionRepository.java
@@ -1,6 +1,0 @@
-package mmm.emopic.app.domain.emotion;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PhotoEmotionRepository extends JpaRepository<PhotoEmotion,Long> {
-}
