+++ emopic/src/main/java/mmm/emopic/app/domain/emotion/repository/EmotionRepository.java
@@ -1,5 +1,6 @@
-package mmm.emopic.app.domain.emotion;
+package mmm.emopic.app.domain.emotion.repository;
 
+import mmm.emopic.app.domain.emotion.Emotion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
